backend/requirements.txt
Django==2.2.19
pytz==2021.3
sqlparse==0.4.2
Pillow==8.3.2
djangorestframework==3.12.4
djoser==2.1.0